+++ sdk-ts/src/index.ts
@@ -10,4 +10,5 @@ export * from "./types.js";
 export * from "./encoding.js";
 export * from "./identity.js";
 export * from "./client.js";
+export * from "./ids.js";
 export * as builders from "./builders.js";